[CO] New as_dict for MeshVTK
