+++ pyleecan/Classes/Class_Dict.json
@@ -5744,7 +5744,8 @@
             "get_normals",
             "get_surf",
             "get_cell_area",
-            "convert"
+            "convert",
+            "as_dict"
         ],
         "mother": "Mesh",
         "name": "MeshVTK",
