+++ pyleecan/Generator/ClassesRef/Mesh/MeshVTK.csv
@@ -6,5 +6,5 @@ path,,Path where the mesh is stored,,str,,,,,,,get_normals,,,
 name,,Name of the mesh file,,str,mesh,,,,,,get_surf,,,
 surf,,Pyvista object of the outer surface,,pyvista.core.pointset.PolyData,None,,,,,,get_cell_area,,,
 is_vtk_surf,,Save the surface mesh in a vtk file,,bool,False,,,,,,convert,,,
-surf_path,,Path where the outer surface is stored,,str,,,,,,,,,,
+surf_path,,Path where the outer surface is stored,,str,,,,,,,as_dict,,,
 surf_name,,Name of the outer surface file,,str,surf,,,,,,,,,
